Change back to 2 pages (undo @omgmog’s fine work)
I need the two of these to be separate; I’m building it for the 
orientation one (which should be really simple and no code) -
the pointer version is to describe how the stores work, but I’m not
going to go through the code for that one.

Also, I refresh browser windows more than I `git status` - so it’s 
handy to not have to press a button to make it work.